+++ src/use-async-task-axios.d.ts
@@ -2,6 +2,11 @@ import { AxiosRequestConfig } from 'axios';
 // eslint-disable-next-line import/named
 import { AsyncTask } from './index';
 
+export const useMemoSafe: <T>(
+  creator: () => T,
+  inputs: unknown[],
+) => T;
+
 export type UseAsyncTaskAxios = <Result>(
   config: AxiosRequestConfig,
 ) => AsyncTask<Result>;