+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Do not use useMemo/useCallback as a semantic guarantee
+- A naive useMemoSafe for axios config
 
 ## [0.7.0] - 2019-01-22
 ### Added